+++ packages/plugins/auto-doc/src/ui/components/views/template-overview/template-overview.svelte
@@ -35,7 +35,7 @@ function fetchTemplates() {
 }
 
 function navigateToCreateTemplate() {
-	navigate({ view: View.Create })
+	navigate({ view: View.Create, templateId: null })
 }
 
 async function onImportTemplate(file: File) {