+++ packages/plugins/auto-doc/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.18.12] - 2026-04-20
 ### Added
 - Invalidities report for the signal list as a dialog
+### Fixed
+- creating "new" template reopening the latest used template
 
 ## [1.18.11] - 2026-02-16
 ### Fixed